Add files via upload
Data mining is like finding hidden treasure in a big pile of information. In this activity, i've analyze the accident data(dataframe) to find patterns, like which weather conditions cause more accidents or which roads are the most dangerous. By studying these patterns, we can help make roads safer and prevent future accidents. It's like being a detective, but for data.